+++ lib/metrics/blacklist.go
@@ -43,6 +43,10 @@ metric target.
 Blacklists can either be path prefixes or regular expression patterns, if either
 a path prefix or regular expression matches a metric path it will be excluded.
 
+Metrics must be matched using dot notation even if the chosen output uses a
+different form. For example, the path would be 'foo.bar' rather than 'foo_bar'
+even when sending metrics to Prometheus.
+
 The ` + "`prefix`" + ` field in a metrics config is ignored by this type. Please
 configure a prefix at the child level.
 
@@ -58,6 +62,22 @@ An entry in the ` + "`patterns`" + ` field will be parsed as an RE2 regular
 expression and tested against each metric path. This can be used, for example,
 to allow none of the latency based metrics with the pattern
 ` + "`.*\\.latency`" + `.`,
+		sanitiseConfigFunc: func(conf Config) (interface{}, error) {
+			var childSanit interface{}
+			var err error
+			if conf.Blacklist.Child != nil {
+				if childSanit, err = SanitiseConfig(*conf.Blacklist.Child); err != nil {
+					return nil, err
+				}
+			} else {
+				childSanit = struct{}{}
+			}
+			return map[string]interface{}{
+				"paths":    conf.Blacklist.Paths,
+				"patterns": conf.Blacklist.Patterns,
+				"child":    childSanit,
+			}, nil
+		},
 	}
 }